refactor(ax): split Herdr queue wake reminder pass
